Server/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==5.0.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
werkzeug==3.1.3
PyJWT==2.10.1
